requirements 2.txt
flask
flask_sqlalchemy
flask_login
flask_migrate
python-dotenv
stripe>=10
psycopg[binary]>=3.2
gunicorn
